LPS-196947 Inline OpenNLPDocumentAssetAutoTagger methods into OpenNLPDocumentAssetAutoTagProvider
